Resolved Merge Conflicts
