+++ dags/daily_weather_model_development_dag .py
@@ -5,16 +5,16 @@ from datetime import datetime, timedelta
 import pandas as pd
 import numpy as np
 import logging
+import json 
 import os
 import pickle
 from utils import save_model_to_gcs, read_data_from_gcs
 from daily_model_training import (
     process_data,
-    train_model,
+    run_model_training,
 )
-from daily_model_validation import evaluate_and_visualize_model
-from daily_model_bias_detection import calculate_metrics_for_features, bin_column
-from daily_model_sensitivity import analyze_model_sensitivity
+from daily_model_bias_detection import run_bias_detection_workflow, plot_bias_metrics
+from daily_model_sensitivity import perform_feature_importance_analysis, perform_hyperparameter_sensitivity_analysis
 from constants import *
 
 # Configure logging
@@ -32,8 +32,8 @@ default_args = {
 }
 
 # Define the DAG
-dag = DAG(
-    'daily_weather_model_development_pipeline',
+dag3 = DAG(
+    'daily_weather_model_development_pipeline_v2',
     default_args=default_args,
     description='DAG for running the model development pipeline on daily weather data',
     schedule_interval=None,
@@ -67,20 +67,26 @@ def notify_failure(context):
     failure_email.execute(context=context)
 
 # Task to load data from GCS
-def load_data_task(**kwargs):
+def daily_load_data_task(**kwargs):
     logging.info("Loading data from GCS")
     data = read_data_from_gcs(BUCKET_NAME, ENGINEERED_DAILY_DATA_PATH)
-    data_path = os.path.join(TEMP_DIR, "raw_data.pkl")
+    data_path = os.path.join(TEMP_DIR, "engineered_data.pkl")
+
+    logging.info(f"Saving loaded data to {data_path}")
     with open(data_path, "wb") as f:
         pickle.dump(data, f)
-    kwargs['ti'].xcom_push(key='raw_data_path', value=data_path)
+
+    if not os.path.exists(data_path):
+        raise FileNotFoundError(f"Data file could not be created at {data_path}")
+
+    kwargs['ti'].xcom_push(key='engineered_data_path', value=data_path)
     logging.info("Data loading task completed.")
 
 # Task to process data
-def process_data_task(**kwargs):
+def daily_process_data_task(**kwargs):
     logging.info("Processing data")
 
-    raw_data_path = kwargs['ti'].xcom_pull(key='raw_data_path', task_ids='load_data')
+    raw_data_path = kwargs['ti'].xcom_pull(key='engineered_data_path', task_ids='load_data')
     with open(raw_data_path, "rb") as f:
         raw_data = pickle.load(f)
 
@@ -89,10 +95,10 @@ def process_data_task(**kwargs):
         'showers_sum', 'daylight_duration', 'precipitation_sum', 'temperature_range',
         'diurnal_temp_range', 'precipitation_intensity'
     ]
-    targets = ['apparent_temperature_max', 'sunshine_duration', 'snowfall_sum', 
-               'precipitation_sum', 'precipitation_hours']
+    targets = ['apparent_temperature_max', 'sunshine_duration', 'snowfall_sum']
 
-    processed_data, scaler_features, target_scalers = process_data(raw_data, features, targets)
+    # Remove scalers from process_data function
+    processed_data, scaler_features, _ = process_data(raw_data, features, targets)
 
     for target, splits in processed_data.items():
         for split, data_split in splits.items():
@@ -112,283 +118,227 @@ def process_data_task(**kwargs):
 
             kwargs['ti'].xcom_push(key=f"{target}_{split}_path", value=split_path)
 
-    for target, scaler in target_scalers.items():
-        scaler_path = os.path.join(TEMP_DIR, f"{target}_scaler.pkl")
-        with open(scaler_path, "wb") as f:
-            pickle.dump(scaler, f)
-        kwargs['ti'].xcom_push(key=f"{target}_scaler_path", value=scaler_path)
-
+    # Remove handling of target scalers entirely
     logging.info("Data processing task completed.")
-
     
-# Task to train model
-def train_model_task(**kwargs):
-    logging.info("Training models for all targets")
-    targets = ['apparent_temperature_max', 'sunshine_duration', 'snowfall_sum', 
-               'precipitation_sum', 'precipitation_hours']
+def daily_train_model_task(**kwargs):
+    logging.info("Starting model training with hyperparameter optimization")
 
-    for target in targets:
-        # Pull data paths from XCom
-        X_train_path = kwargs['ti'].xcom_pull(key=f"{target}_X_train_path", task_ids='process_data')
-        X_val_path = kwargs['ti'].xcom_pull(key=f"{target}_X_val_path", task_ids='process_data')
-        y_train_path = kwargs['ti'].xcom_pull(key=f"{target}_y_train_path", task_ids='process_data')
-        y_val_path = kwargs['ti'].xcom_pull(key=f"{target}_y_val_path", task_ids='process_data')
+    # Retrieve the data path from XCom
+    data_path = kwargs['ti'].xcom_pull(key='engineered_data_path', task_ids='load_data')
 
-        with open(X_train_path, "rb") as f:
-            X_train = pickle.load(f)
-        with open(X_val_path, "rb") as f:
-            X_val = pickle.load(f)
-        with open(y_train_path, "rb") as f:
-            y_train = pickle.load(f)
-        with open(y_val_path, "rb") as f:
-            y_val = pickle.load(f)
+    if not data_path:
+        raise ValueError("No data path found in XCom for 'engineered_data_path'. Ensure 'load_data_task' pushed the correct path.")
 
-        model = train_model(X_train, X_val, y_train, y_val)
+    logging.info(f"Data path retrieved from XCom: {data_path}")
 
-        # Save the model
-        model_path = os.path.join(TEMP_DIR, f"{target}_model.pkl")
-        with open(model_path, "wb") as f:
-            pickle.dump(model, f)
-        kwargs['ti'].xcom_push(key=f"{target}_model_path", value=model_path)
+    if not os.path.exists(data_path):
+        raise FileNotFoundError(f"Data file not found at path: {data_path}")
 
-    logging.info("Model training completed.")
+    # Define the output path for the combined models file
+    combined_output_file = os.path.join(TEMP_DIR, "all_models.pkl")
 
-# Task to validate models and generate plots
-def validate_models_task(**kwargs):
-    logging.info("Validating models and generating visualizations")
+    # Pass the data to the model_training function
+    models_file_path = run_model_training(data_path, combined_output_file, task_instance=kwargs['ti'])
 
-    targets = ['apparent_temperature_max', 'sunshine_duration', 'snowfall_sum', 
-               'precipitation_sum', 'precipitation_hours']
+    # Push the path of the combined models file to XCom
+    kwargs['ti'].xcom_push(key='models_file_path', value=models_file_path)
+    logging.info(f"Model training task completed successfully. Models saved at {models_file_path}")
+
+
+# Task for bias detection (remove dependency on scalers)
+def daily_bias_detection_task(**kwargs):
+    logging.info("Starting bias detection")
+
+    # Load processed data path
+    data_path = kwargs['ti'].xcom_pull(key='engineered_data_path', task_ids='load_data')
+    with open(data_path, "rb") as f:
+        data = pickle.load(f)
     
-    # Retrieve processed data and scalers from XCom for each target
-    processed_data = {}
-    scalers = {}
-
-    for target in targets:
-        processed_data[target] = {}
-        
-        # Retrieve processed data splits for the target from XCom
-        for split in ['X_val', 'X_test', 'y_val', 'y_test']:
-            split_path = kwargs['ti'].xcom_pull(key=f"{target}_{split}_path", task_ids='process_data')
-            with open(split_path, "rb") as f:
-                processed_data[target][split] = pickle.load(f)
-        
-        # Retrieve scaler for the target
-        scaler_path = kwargs['ti'].xcom_pull(key=f"{target}_scaler_path", task_ids='process_data')
-        with open(scaler_path, "rb") as f:
-            scalers[target] = pickle.load(f)
-
-        # Retrieve the model for the target
-        model_path = kwargs['ti'].xcom_pull(key=f"{target}_model_path", task_ids='train_model')
-        with open(model_path, "rb") as f:
-            model = pickle.load(f)
-        
-        # Evaluate and visualize the model for the current target
-        evaluate_and_visualize_model(
-            model=model,
-            processed_data=processed_data,  # Pass all processed data; function uses only required parts
-            targets=[target],               # Single target to match function expectations
-            scalers=scalers,                # Pass all scalers; function uses only required parts
-            bucket_name=BUCKET_NAME,
-            WEATHER_DATA_PLOTS_PATH=WEATHER_DATA_PLOTS_PATH
-        )
-        logging.info(f"Model validation and visualization for target {target} completed.")
-    
-    logging.info("All models validated and visualized successfully.")
-
-# Task to detect bias
-def bias_detection_task(**kwargs):
-    logging.info("Performing bias detection")
-
-    slicing_features = ['temperature_2m_max', 'precipitation_sum']  # Features to slice by
-    targets = ['apparent_temperature_max', 'sunshine_duration', 'snowfall_sum',
-               'precipitation_sum', 'precipitation_hours']
+    # Load models
     models = {}
-    scalers = {}
-
-    # Load models and scalers
-    for target in targets:
+    for target in ['apparent_temperature_max', 'sunshine_duration', 'snowfall_sum']:
         model_path = kwargs['ti'].xcom_pull(key=f"{target}_model_path", task_ids='train_model')
         with open(model_path, "rb") as f:
             models[target] = pickle.load(f)
-
-        scaler_path = kwargs['ti'].xcom_pull(key=f"{target}_scaler_path", task_ids='process_data')
-        with open(scaler_path, "rb") as f:
-            scalers[target] = pickle.load(f)
-
-    # Prepare processed data
-    for target in targets:
-        X_test_path = kwargs['ti'].xcom_pull(key=f"{target}_X_test_path", task_ids='process_data')
-        y_test_path = kwargs['ti'].xcom_pull(key=f"{target}_y_test_path", task_ids='process_data')
-
-        with open(X_test_path, "rb") as f:
-            X_test = pickle.load(f)
-        with open(y_test_path, "rb") as f:
-            y_test = pickle.load(f)
-
-        # Recreate DataFrame with all features and target
-        features = [
-            'temperature_2m_max', 'temperature_2m_min', 'apparent_temperature_min', 'rain_sum',
-            'showers_sum', 'daylight_duration', 'precipitation_sum', 'temperature_range',
-            'diurnal_temp_range', 'precipitation_intensity'
-        ]
-        data = pd.DataFrame(X_test, columns=features)
-        data[target] = y_test  # Add target column to the DataFrame
-
-        # Ensure slicing features exist in the dataset
-        missing_features = [feature for feature in slicing_features if feature not in data.columns]
-        if missing_features:
-            logging.error(f"Slicing features {missing_features} are missing in the processed data for {target}")
-            continue
-
-        logging.info(f"Data columns for target {target}: {list(data.columns)}")
-
-        # Perform bias detection
-        try:
-            metrics_and_interpretations = calculate_metrics_for_features(
-                data=data,
-                features=slicing_features,
-                models={target: models[target]},
-                scalers={target: scalers[target]},
-                targets=[target],  # Pass the current target
-                threshold_ratio=0.1
-            )
-
-            # Log interpretations
-            for interpretation in metrics_and_interpretations["interpretations"]:
-                logging.info(interpretation)
-
-        except ValueError as e:
-            logging.error(f"Bias detection failed for target {target}: {str(e)}")
-            logging.warning(f"Skipping bias detection for {target} and continuing the pipeline.")
-
-    logging.info("Bias detection task completed.")
     
-# Task to perform model sensitivity analysis
-def model_sensitivity_task(**kwargs):
-    logging.info("Performing model sensitivity analysis using SHAP")
+    features = [
+    'temperature_2m_max', 'temperature_2m_min', 'apparent_temperature_min', 'rain_sum',
+    'showers_sum', 'daylight_duration', 'precipitation_sum', 'temperature_range',
+    'diurnal_temp_range', 'precipitation_intensity'
+    ]
+    
+    # Define slicing features and their configurations
+    slicing_definitions = {
+        'temperature_2m_max_bin': {'column': 'temperature_2m_max', 'bins': 4, 'labels': ['cold', 'cool', 'warm', 'hot']},
+        'precipitation_sum_bin': {'column': 'precipitation_sum', 'bins': 4, 'labels': ['low', 'medium-low', 'medium-high', 'high']}
+        }
 
-    targets = ['apparent_temperature_max', 'sunshine_duration', 'snowfall_sum',
-               'precipitation_sum', 'precipitation_hours']
-    models = {}
-    processed_data = {}
+    # List of target variables
+    targets = ['apparent_temperature_max', 'sunshine_duration', 'snowfall_sum']
 
-    # Load models and processed data
+    # Dictionary to store metrics for all targets
+    all_metrics = {}
+
+    # Loop through each target
     for target in targets:
-        # Load the model
-        model_path = kwargs['ti'].xcom_pull(key=f"{target}_model_path", task_ids='train_model')
-        with open(model_path, "rb") as f:
-            models[target] = pickle.load(f)
+        logging.info(f"Starting bias detection workflow for target: {target}")
 
-        # Load the processed test data
-        X_test_path = kwargs['ti'].xcom_pull(key=f"{target}_X_test_path", task_ids='process_data')
-        with open(X_test_path, "rb") as f:
-            X_test = pickle.load(f)
+        # Run bias detection workflow for the current target
+        metrics = run_bias_detection_workflow(data, features, target, models[target], slicing_definitions)
 
-        # Store test data for SHAP analysis
-        processed_data[target] = {"X_test": X_test}
+        # Store the metrics for the current target
+        all_metrics[target] = metrics
 
-    # Perform sensitivity analysis for all targets
+    # Print and visualize bias detection metrics for all targets
+    for target, metrics in all_metrics.items():
+        print(f"\nBias Metrics for {target}:")
+        for slicing_feature, (overall, by_group) in metrics.items():
+            print(f"  Slicing Feature: {slicing_feature}")
+            print(f"    Overall RMSE: {overall}")
+            print(f"    By Group RMSE: {by_group}")
+            plot_bias_metrics(metrics, slicing_features=[slicing_feature])
+
+
+    logging.info("Bias detection task completed successfully.")
+
+    
+def daily_model_sensitivity_task(bucket_name=BUCKET_NAME, **kwargs):
+    logging.info("Performing model sensitivity analysis")
+
+    targets = ['apparent_temperature_max', 'sunshine_duration', 'snowfall_sum']
     features = [
         'temperature_2m_max', 'temperature_2m_min', 'apparent_temperature_min', 'rain_sum',
         'showers_sum', 'daylight_duration', 'precipitation_sum', 'temperature_range',
         'diurnal_temp_range', 'precipitation_intensity'
     ]
-    analyze_model_sensitivity(
-        models=models,
-        processed_data=processed_data,
-        features=features,
-        targets=targets,
-        bucket_name=BUCKET_NAME,
-        weather_data_plots_path=WEATHER_DATA_PLOTS_PATH
-    )
-    logging.info("Model sensitivity analysis task completed.")
-
-    
-# Task to save model
-def save_model_task(**kwargs):
-    logging.info("Saving models to GCS")
-    targets = ['apparent_temperature_max', 'sunshine_duration', 'snowfall_sum', 
-               'precipitation_sum', 'precipitation_hours']
+    models = {}
+    processed_data = {}
 
     for target in targets:
+        # Load model
         model_path = kwargs['ti'].xcom_pull(key=f"{target}_model_path", task_ids='train_model')
+        if not model_path or not os.path.exists(model_path):
+            logging.error(f"Model path for {target} is missing or invalid: {model_path}")
+            continue
         with open(model_path, "rb") as f:
-            model = pickle.load(f)
+            model_data = pickle.load(f)
+            models[target] = model_data.get("model")
+            if models[target] is None:
+                logging.error(f"Failed to extract model for {target}")
+                continue
 
-        # Save the model to GCS
-        save_model_to_gcs(
-            model=model,
-            bucket_name=BUCKET_NAME,
-            file_name=f"assets/daily_models/{target}_model.pkl"
-        )
-    logging.info("All models saved to GCS.")
+        # Load test data
+        X_test_path = kwargs['ti'].xcom_pull(key=f"{target}_X_test_path", task_ids='process_data')
+        if not X_test_path or not os.path.exists(X_test_path):
+            logging.error(f"Test data for {target} is missing or invalid: {X_test_path}")
+            continue
+        with open(X_test_path, "rb") as f:
+            X_test = pickle.load(f)[features]  # Filter only relevant features
+
+        # Perform feature importance analysis
+        try:
+            perform_feature_importance_analysis(
+                model=models[target],
+                X_test=X_test,
+                feature_names=features,
+                target=target,
+                bucket_name=bucket_name,
+            )
+        except Exception as e:
+            logging.error(f"Feature importance analysis failed for {target}: {e}")
+
+        # Load trials and perform hyperparameter sensitivity analysis
+        trial_path = kwargs['ti'].xcom_pull(key=f"{target}_trials_path", task_ids='train_model')
+        if not trial_path or not os.path.exists(trial_path):
+            logging.error(f"Trials data for {target} is missing or invalid: {trial_path}")
+            continue
+        with open(trial_path, "rb") as f:
+            trials = pickle.load(f)
+
+        try:
+            perform_hyperparameter_sensitivity_analysis(trials, target, bucket_name)
+        except Exception as e:
+            logging.error(f"Hyperparameter sensitivity analysis failed for {target}: {e}")
+
+    logging.info("Model sensitivity analysis task completed successfully.")
+    
+# Task to save best models to GCS
+def save_models_to_gcs_task(**kwargs):
+    logging.info("Saving best models to GCS")
+
+    # Retrieve the best model paths from XCom
+    best_model_paths = kwargs['ti'].xcom_pull(key='models_file_path', task_ids='train_model')
+
+    # Define the GCS path for the model
+    gcs_path = f"assets/daily_models/daily_best_model.pkl"
+
+    # Save the model to GCS
+    save_model_to_gcs(
+    model=best_model_paths,
+    bucket_name=BUCKET_NAME,
+    file_name= "daily_best_model"
+    )
+    logging.info(f"Best mode saved to GCS at {gcs_path}")
 
 # Define Airflow tasks
 load_data_operator = PythonOperator(
     task_id='load_data',
-    python_callable=load_data_task,
+    python_callable=daily_load_data_task,
     provide_context=True,
     on_failure_callback=notify_failure,
-    dag=dag
+    dag=dag3
 )
 
 process_data_operator = PythonOperator(
     task_id='process_data',
-    python_callable=process_data_task,
+    python_callable=daily_process_data_task,
     provide_context=True,
     on_failure_callback=notify_failure,
-    dag=dag
+    dag=dag3
 )
 
 train_model_operator = PythonOperator(
     task_id='train_model',
-    python_callable=train_model_task,
+    python_callable=daily_train_model_task,
     provide_context=True,
     on_failure_callback=notify_failure,
-    dag=dag
-)
-
-validate_models_operator = PythonOperator(
-    task_id='validate_models',
-    python_callable=validate_models_task,
-    provide_context=True,
-    on_failure_callback=notify_failure,
-    dag=dag
+    execution_timeout=timedelta(minutes=10),
+    dag=dag3
 )
 
 bias_detection_operator = PythonOperator(
     task_id='bias_detection',
-    python_callable=bias_detection_task,
+    python_callable=daily_bias_detection_task,
     provide_context=True,
     on_failure_callback=notify_failure,
-    dag=dag
+    dag=dag3
 )
 
 # Define the task in the DAG
 model_sensitivity_operator = PythonOperator(
     task_id='model_sensitivity',
-    python_callable=model_sensitivity_task,
+    python_callable=daily_model_sensitivity_task,
     provide_context=True,
     on_failure_callback=notify_failure,
-    dag=dag
+    dag=dag3
 )
 
-save_model_operator = PythonOperator(
-    task_id='save_model',
-    python_callable=save_model_task,
+save_best_model_operator = PythonOperator(
+    task_id='save_best_model',
+    python_callable=save_models_to_gcs_task,
     provide_context=True,
     on_failure_callback=notify_failure,
-    dag=dag
+    dag=dag3
 )
 
 email_notification_task = EmailOperator(
     task_id='send_email_notification',
-    to='darshan.webjaguar@gmail.com',
+    to='keshiarun01@gmail.com',
     subject='Daily Model Development Pipeline Completed Successfully',
     html_content='<p>The Daily Model Development Pipeline DAG has completed successfully.</p>',
-    dag=dag
+    dag=dag3
 )
 
 # Set task dependencies
-load_data_operator >> process_data_operator >> train_model_operator >> validate_models_operator >> bias_detection_operator >> model_sensitivity_operator >> save_model_operator >> email_notification_task
+load_data_operator >> process_data_operator >> train_model_operator >> bias_detection_operator >> model_sensitivity_operator >> save_best_model_operator >> email_notification_task
